Start books backend
