--- kevin-backend-userauth/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-
-keys.js
