--- kevin-backend-userauth/app.js
@@ -1,26 +1,0 @@
-const express = require("express");
-const app = express();
-const mongoose = require('mongoose');
-const db = require('./config/keys').mongoURI;
-const bodyParser = require('body-parser');
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
-
-app.get("/", (req, res) => res.send("Hello"));
-
-const port = process.env.PORT || 5000;
-
-app.listen(port, () => console.log(`server is running on port ${port}`));
-
-mongoose
-    .connect(db, { useNewUrlParser: true })
-    .then(() => console.log("Connected to MongoDB successfully"))
-    .catch(err => console.log(err));
-
-const users = require("./routes/api/users");
-const books = require("./routes/api/books");
-
-app.use("/api/users", users);
-app.use("/api/books", books);
-
-
