--- kevin-backend-userauth/models/Book.js
@@ -1,27 +1,0 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-
-const BookSchema = new Schema({
-    title: {
-        type: String,
-        required: true
-    },
-    author: {
-        type: String,
-        required: true
-    },
-    genre: {
-        type: String,
-        required: true
-    },
-    length: {
-        type: Number,
-        required: true
-    },
-    date: {
-        type: Date,
-        required: true
-    }
-});
-
-module.exports = Book = mongoose.model('books', BookSchema);
