--- kevin-backend-userauth/models/User.js
@@ -1,23 +1,0 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-
-const UserSchema = new Schema({
-    handle: {
-        type: String,
-        required: true
-    },
-    email: {
-        type: String,
-        required: true
-    },
-    password: {
-        type: String,
-        required: true
-    },
-    date: {
-        type: Date,
-        default: Date.now
-    }
-})
-
-module.exports = User = mongoose.model('users', UserSchema);
