--- kevin-backend-userauth/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "alexandria",
-  "version": "1.0.0",
-  "description": "# Background and Overview Alexandria is a book app.  It aims to help readers keep track of their books, and to help them discover new reads.",
-  "main": "app.js",
-  "scripts": {
-    "start": "node app.js",
-    "server:debug": "nodemon --inspect app.js",
-    "server": "nodemon app.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zkevinbai/Alexandria.git"
-  },
-  "author": "kevin bai",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/zkevinbai/Alexandria/issues"
-  },
-  "homepage": "https://github.com/zkevinbai/Alexandria#readme",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.18.3",
-    "eslint": "^5.15.1",
-    "express": "^4.16.4",
-    "jsonwebtoken": "^8.5.0",
-    "mongoose": "^5.4.18",
-    "nodemon": "^1.18.10",
-    "passport": "^0.4.0",
-    "passport-jwt": "^4.0.0",
-    "validator": "^10.11.0"
-  }
-}
