--- kevin-backend-userauth/routes/api/books.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get("/test", (req, res) => res.json({
-    msg: "this is the books route"
-}));
-
-module.exports = router;
